[Refactor][Philippe] Modularized Styles, Modals and API Calls of Edit Account.
